requirements.txt
streamlit
pillow
pandas
google-generativeai
python-dotenv
